0909 han OrderCreate 페이지 Join 테이블 Insert JSQL 작업

+++ src/main/java/com/project/erpre/model/Category.java
@@ -19,7 +19,7 @@ public class Category {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "category_no")
-    private Long categoryNo;
+    private Integer categoryNo;
 
     @Column(name = "category_level", nullable = false)
     private Integer categoryLevel;
